requirements.txt
streamlit
langchain
langchain-openai
python-dotenv
pypdf
faiss-cpu
